% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.robmedfit}
\alias{print.robmedfit}
\title{Print a robmedfit object}
\usage{
\method{print}{robmedfit}(x, ...)
}
\arguments{
\item{x}{A \code{robmedfit} object returned by \code{\link[=robustmediate]{robustmediate()}}.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print a robmedfit object
}
